fix: remove initialBytePos truncation, fix use-after-free, sign-extension, and hardcoded key length
- PageKind: remove `& 0xFF` masks on initialBytePos in BI_NODE, SPAN_NODE, MULTI_NODE
  deserialization — value already read as Short.toUnsignedInt, mask truncated keys > 255 bytes
- HOTLeafPage: reorder ensureMutableSlotMemory() to release old memory AFTER assigning new
  slotMemory and releaser, preventing use-after-free if allocation succeeds but later ops fail
- HOTIndirectPage: change old createMultiNode(byte) overload parameter to int, preventing
  sign-extension for byte positions > 127
- HOTLongIndexWriter: use keySerializer return value for key length in remove() instead of
  hardcoded 8

+++ bundles/sirix-core/src/main/java/io/sirix/page/HOTIndirectPage.java
@@ -260,7 +260,7 @@ public final class HOTIndirectPage implements Page {
    * @param children array of child references
    * @return new MultiNode
    */
-  public static HOTIndirectPage createMultiNode(long pageKey, int revision, byte discriminativeByte, byte[] childIndex,
+  public static HOTIndirectPage createMultiNode(long pageKey, int revision, int discriminativeByte, byte[] childIndex,
       PageReference[] children) {
     if (children.length < 1 || children.length > MAX_NODE_ENTRIES) {
       throw new IllegalArgumentException("MultiNode must have 1-" + MAX_NODE_ENTRIES + " children, got: "
